Improve C# Client
The C# Client now uses an array of 7 bytes instead of one big UInt64. As a result, the button state dictionary has UInt16 as it's value type and there is some less typecasting shenanigans. The Joystick axes now use the map function and can report a value of -127. This generally improves everything and makes it a bit easier to understand at a glance.