Create ByteSequenceExplanation.txt
This file serves as an explanation for the format used to communicate the controller state to the Arduino. It is a much more readable, well structured and correctly spelled alternative to the stream-of-thought mess that lies in ControllerMapping.txt, which got to it's current state by being a convenient place to jot down notes during development as I was planning the format.

I have also added a message to ControllerMapping advising people read this instead.